Refactored screenshot, startRecording, and stopRecording into their own files. Added the overwrite option for startRecording.
